Move app polling methods from docker_app_runner to app_examiner.
[Finishes #91092888]
