+++ ltc/app_examiner/app_examiner_test.go
@@ -496,4 +496,81 @@ var _ = Describe("AppRunner", func() {
 			})
 		})
 	})
+
+	Describe("NumOfRunningAppInstances", func() {
+		It("returns the number of running instances for a given app guid", func() {
+			actualLrpsResponse := []receptor.ActualLRPResponse{
+				receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateRunning, Index: 1},
+				receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateRunning, Index: 2},
+				receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateClaimed, Index: 3},
+			}
+			fakeReceptorClient.ActualLRPsByProcessGuidReturns(actualLrpsResponse, nil)
+
+			count, placementError, err := appExaminer.RunningAppInstancesInfo("americano-app")
+			Expect(placementError).To(BeFalse())
+			Expect(err).ToNot(HaveOccurred())
+			Expect(count).To(Equal(2))
+
+			Expect(fakeReceptorClient.ActualLRPsByProcessGuidCallCount()).To(Equal(1))
+			Expect(fakeReceptorClient.ActualLRPsByProcessGuidArgsForCall(0)).To(Equal("americano-app"))
+		})
+
+		It("returns errors from the receptor", func() {
+			receptorError := errors.New("receptor did not like that requeset")
+			fakeReceptorClient.ActualLRPsByProcessGuidReturns([]receptor.ActualLRPResponse{}, receptorError)
+
+			_, _, err := appExaminer.RunningAppInstancesInfo("nescafe-app")
+
+			Expect(err).To(HaveOccurred())
+			Expect(err).To(Equal(receptorError))
+		})
+
+		Context("when there are placement errors on an instance", func() {
+			It("returns true for placementError", func() {
+				actualLrpsResponse := []receptor.ActualLRPResponse{
+					receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateRunning, Index: 1},
+					receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateUnclaimed, Index: 2, PlacementError: "could not place!"},
+					receptor.ActualLRPResponse{ProcessGuid: "americano-app", State: receptor.ActualLRPStateRunning, Index: 3},
+				}
+				fakeReceptorClient.ActualLRPsByProcessGuidReturns(actualLrpsResponse, nil)
+
+				count, placementError, err := appExaminer.RunningAppInstancesInfo("americano-app")
+				Expect(placementError).To(BeTrue())
+				Expect(err).ToNot(HaveOccurred())
+				Expect(count).To(Equal(2))
+			})
+		})
+	})
+
+	Describe("AppExists", func() {
+		It("returns true if the docker app exists", func() {
+			actualLRPs := []receptor.ActualLRPResponse{receptor.ActualLRPResponse{ProcessGuid: "americano-app"}}
+			fakeReceptorClient.ActualLRPsReturns(actualLRPs, nil)
+
+			exists, err := appExaminer.AppExists("americano-app")
+			Expect(err).ToNot(HaveOccurred())
+			Expect(exists).To(BeTrue())
+		})
+
+		It("returns false if the docker app does not exist", func() {
+			actualLRPs := []receptor.ActualLRPResponse{}
+			fakeReceptorClient.ActualLRPsReturns(actualLRPs, nil)
+
+			exists, err := appExaminer.AppExists("americano-app")
+			Expect(err).ToNot(HaveOccurred())
+			Expect(exists).To(BeFalse())
+		})
+
+		Describe("returning errors from the receptor", func() {
+			It("returns errors fetching the status", func() {
+				actualLRPs := []receptor.ActualLRPResponse{}
+				fakeReceptorClient.ActualLRPsReturns(actualLRPs, errors.New("Something Bad"))
+
+				exists, err := appExaminer.AppExists("americano-app")
+				Expect(err).To(HaveOccurred())
+				Expect(err.Error()).To(Equal("Something Bad"))
+				Expect(exists).To(BeFalse())
+			})
+		})
+	})
 })
