+++ ltc/app_examiner/fake_app_examiner/fake_app_examiner.go
@@ -31,6 +31,25 @@ type FakeAppExaminer struct {
 		result1 app_examiner.AppInfo
 		result2 error
 	}
+	AppExistsStub        func(name string) (bool, error)
+	appExistsMutex       sync.RWMutex
+	appExistsArgsForCall []struct {
+		name string
+	}
+	appExistsReturns struct {
+		result1 bool
+		result2 error
+	}
+	RunningAppInstancesInfoStub        func(name string) (int, bool, error)
+	runningAppInstancesInfoMutex       sync.RWMutex
+	runningAppInstancesInfoArgsForCall []struct {
+		name string
+	}
+	runningAppInstancesInfoReturns struct {
+		result1 int
+		result2 bool
+		result3 error
+	}
 }
 
 func (fake *FakeAppExaminer) ListApps() ([]app_examiner.AppInfo, error) {
@@ -114,6 +133,73 @@ func (fake *FakeAppExaminer) AppStatusReturns(result1 app_examiner.AppInfo, resu
 		result1 app_examiner.AppInfo
 		result2 error
 	}{result1, result2}
+}
+
+func (fake *FakeAppExaminer) AppExists(name string) (bool, error) {
+	fake.appExistsMutex.Lock()
+	fake.appExistsArgsForCall = append(fake.appExistsArgsForCall, struct {
+		name string
+	}{name})
+	fake.appExistsMutex.Unlock()
+	if fake.AppExistsStub != nil {
+		return fake.AppExistsStub(name)
+	} else {
+		return fake.appExistsReturns.result1, fake.appExistsReturns.result2
+	}
+}
+
+func (fake *FakeAppExaminer) AppExistsCallCount() int {
+	fake.appExistsMutex.RLock()
+	defer fake.appExistsMutex.RUnlock()
+	return len(fake.appExistsArgsForCall)
+}
+
+func (fake *FakeAppExaminer) AppExistsArgsForCall(i int) string {
+	fake.appExistsMutex.RLock()
+	defer fake.appExistsMutex.RUnlock()
+	return fake.appExistsArgsForCall[i].name
+}
+
+func (fake *FakeAppExaminer) AppExistsReturns(result1 bool, result2 error) {
+	fake.AppExistsStub = nil
+	fake.appExistsReturns = struct {
+		result1 bool
+		result2 error
+	}{result1, result2}
+}
+
+func (fake *FakeAppExaminer) RunningAppInstancesInfo(name string) (int, bool, error) {
+	fake.runningAppInstancesInfoMutex.Lock()
+	fake.runningAppInstancesInfoArgsForCall = append(fake.runningAppInstancesInfoArgsForCall, struct {
+		name string
+	}{name})
+	fake.runningAppInstancesInfoMutex.Unlock()
+	if fake.RunningAppInstancesInfoStub != nil {
+		return fake.RunningAppInstancesInfoStub(name)
+	} else {
+		return fake.runningAppInstancesInfoReturns.result1, fake.runningAppInstancesInfoReturns.result2, fake.runningAppInstancesInfoReturns.result3
+	}
+}
+
+func (fake *FakeAppExaminer) RunningAppInstancesInfoCallCount() int {
+	fake.runningAppInstancesInfoMutex.RLock()
+	defer fake.runningAppInstancesInfoMutex.RUnlock()
+	return len(fake.runningAppInstancesInfoArgsForCall)
+}
+
+func (fake *FakeAppExaminer) RunningAppInstancesInfoArgsForCall(i int) string {
+	fake.runningAppInstancesInfoMutex.RLock()
+	defer fake.runningAppInstancesInfoMutex.RUnlock()
+	return fake.runningAppInstancesInfoArgsForCall[i].name
+}
+
+func (fake *FakeAppExaminer) RunningAppInstancesInfoReturns(result1 int, result2 bool, result3 error) {
+	fake.RunningAppInstancesInfoStub = nil
+	fake.runningAppInstancesInfoReturns = struct {
+		result1 int
+		result2 bool
+		result3 error
+	}{result1, result2, result3}
 }
 
 var _ app_examiner.AppExaminer = new(FakeAppExaminer)
